requirements.txt
Django==4.0.5
djangorestframework==3.13.1
Pillow==9.1.1
requests==2.27.1
gunicorn==20.0.4
drf-yasg==1.20.0
